+++ app/src/main/AndroidManifest.xml
@@ -18,21 +18,21 @@
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="AIzaSyBsO2NUfvIVobYpGAHkUdjrwDkTqs1UvxU" />
-        <activity android:name=".LoginActivity"/>
-        <activity android:name=".HomeActivity"
+        <activity android:name=".view.login.LoginActivity"/>
+        <activity android:name=".view.home.HomeActivity"
             android:windowSoftInputMode="stateHidden">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".PlaceDetailActivity"/>
-        <activity android:name=".ReviewCreateActivity"/>
-        <activity android:name=".MyPlaceActivity"/>
-        <activity android:name=".PlaceRegisterActivity"/>
-        <activity android:name=".MyCourseActivity"/>
-        <activity android:name=".MyReviewActivity"/>
-        <activity android:name=".UserModifyActivity"/>
-        <activity android:name=".SearchAddressActivity"/>
+        <activity android:name=".view.place.PlaceDetailActivity"/>
+        <activity android:name=".view.review.ReviewCreateActivity"/>
+        <activity android:name=".view.mypage.myplace.MyPlaceActivity"/>
+        <activity android:name=".view.mypage.myplace.PlaceRegisterActivity"/>
+        <activity android:name=".view.mypage.mycourse.MyCourseActivity"/>
+        <activity android:name=".view.mypage.myreview.MyReviewActivity"/>
+        <activity android:name=".view.login.UserModifyActivity"/>
+        <activity android:name=".view.mypage.myplace.SearchAddressActivity"/>
     </application>
 </manifest>
