+++ app/src/main/java/com/yeonae/chamelezone/adapter/ImageViewPagerAdapter.kt
@@ -1,13 +1,19 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.adapter
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import kotlinx.android.synthetic.main.slider_image.view.*
 import androidx.viewpager.widget.PagerAdapter
+import com.yeonae.chamelezone.R
 
 class ImageViewPagerAdapter: PagerAdapter() {
-    private val images = intArrayOf(R.drawable.img1, R.drawable.img2, R.drawable.img3, R.drawable.img4)
+    private val images = intArrayOf(
+        R.drawable.img1,
+        R.drawable.img2,
+        R.drawable.img3,
+        R.drawable.img4
+    )
 
     override fun isViewFromObject(view: View, `object`: Any): Boolean {
         return view == `object`
