+++ app/src/main/java/com/yeonae/chamelezone/adapter/PagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.adapter
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
