+++ app/src/main/java/com/yeonae/chamelezone/view/course/CourseTabFragment.kt
@@ -1,13 +1,14 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.course
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.yeonae.chamelezone.R
 
-class LikeTabFragment : Fragment() {
+class CourseTabFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
-        inflater.inflate(R.layout.fragment_like, container, false)
+        inflater.inflate(R.layout.fragment_course, container, false)
 }
