+++ app/src/main/java/com/yeonae/chamelezone/view/login/ConfirmEmailFragment.kt
@@ -1,10 +1,11 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.login
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.yeonae.chamelezone.R
 import kotlinx.android.synthetic.main.fragment_confirm_email.*
 
 class ConfirmEmailFragment : Fragment() {
@@ -30,11 +31,13 @@ class ConfirmEmailFragment : Fragment() {
         val email = arguments!!.getString(userEmail)
         tv_email.text = email
         btn_find_password.setOnClickListener {
-            (activity as LoginActivity).replace(FindPasswordFragment(), true)
+            (activity as LoginActivity).replace(
+                FindPasswordFragment(), true)
         }
 
         btn_login.setOnClickListener {
-            (activity as LoginActivity).replace(LoginFragment(), true)
+            (activity as LoginActivity).replace(
+                LoginFragment(), true)
         }
 
         btn_back.setOnClickListener {
