+++ app/src/main/java/com/yeonae/chamelezone/view/login/FindPasswordFragment.kt
@@ -1,4 +1,4 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.login
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,8 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
+import com.yeonae.chamelezone.AlertDialogFragment
+import com.yeonae.chamelezone.R
 import kotlinx.android.synthetic.main.fragment_find_password.*
 
 class FindPasswordFragment : Fragment() {
@@ -50,9 +52,10 @@ class FindPasswordFragment : Fragment() {
     }
 
     private fun showDialog() {
-        val newFragment = AlertDialogFragment.newInstance(
-            "입력하신 정보는 존재하지 않습니다."
-        )
+        val newFragment =
+            AlertDialogFragment.newInstance(
+                "입력하신 정보는 존재하지 않습니다."
+            )
         newFragment.show(fragmentManager!!, "dialog")
     }
 }
