+++ app/src/main/java/com/yeonae/chamelezone/view/login/JoinFragment.kt
@@ -1,4 +1,4 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.login
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
+import com.yeonae.chamelezone.R
 import kotlinx.android.synthetic.main.fragment_join.*
 import java.util.regex.Pattern
