+++ app/src/main/java/com/yeonae/chamelezone/view/login/LoginActivity.kt
@@ -1,8 +1,9 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.login
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
+import com.yeonae.chamelezone.R
 
 class LoginActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
