+++ app/src/main/java/com/yeonae/chamelezone/view/login/LoginFragment.kt
@@ -1,12 +1,13 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.login
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
+import com.yeonae.chamelezone.AlertDialogFragment
+import com.yeonae.chamelezone.R
 import kotlinx.android.synthetic.main.fragment_login.*
 
 
@@ -62,9 +63,10 @@ class LoginFragment : Fragment() {
     }
 
     private fun showDialog() {
-        val newFragment = AlertDialogFragment.newInstance(
-            "입력하신 정보는 존재하지 않습니다."
-        )
+        val newFragment =
+            AlertDialogFragment.newInstance(
+                "입력하신 정보는 존재하지 않습니다."
+            )
         newFragment.show(fragmentManager!!, "dialog")
     }
