+++ app/src/main/java/com/yeonae/chamelezone/view/login/UserModifyActivity.kt
@@ -1,15 +1,15 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.login
 
 import android.os.Bundle
-import android.widget.ArrayAdapter
 import androidx.appcompat.app.AppCompatActivity
-import kotlinx.android.synthetic.main.activity_my_review.*
+import com.yeonae.chamelezone.R
+import kotlinx.android.synthetic.main.activity_user_modify.*
 
-class MyReviewActivity : AppCompatActivity() {
+class UserModifyActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_my_review)
+        setContentView(R.layout.activity_user_modify)
 
         btn_back.setOnClickListener {
             finish()
