+++ app/src/main/java/com/yeonae/chamelezone/view/map/GpsPermission.kt
@@ -1,4 +1,4 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.map
 
 import android.Manifest
 import android.content.pm.PackageManager
@@ -7,6 +7,7 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
+import com.yeonae.chamelezone.R
 
 class GpsPermission : AppCompatActivity() {
