+++ app/src/main/java/com/yeonae/chamelezone/view/mypage/mycourse/MyCourseActivity.kt
@@ -1,8 +1,8 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.mypage.mycourse
 
 import android.os.Bundle
-import android.widget.ArrayAdapter
 import androidx.appcompat.app.AppCompatActivity
+import com.yeonae.chamelezone.R
 import kotlinx.android.synthetic.main.activity_my_course.*
 
 class MyCourseActivity : AppCompatActivity() {
