+++ app/src/main/java/com/yeonae/chamelezone/view/mypage/myplace/MyPlaceActivity.kt
@@ -1,8 +1,9 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.mypage.myplace
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.yeonae.chamelezone.R
 import kotlinx.android.synthetic.main.activity_my_place.*
 
 class MyPlaceActivity : AppCompatActivity() {
