+++ app/src/main/java/com/yeonae/chamelezone/view/mypage/myreview/MyReviewActivity.kt
@@ -1,15 +1,15 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.mypage.myreview
 
-import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import kotlinx.android.synthetic.main.activity_user_modify.*
+import com.yeonae.chamelezone.R
+import kotlinx.android.synthetic.main.activity_my_review.*
 
-class UserModifyActivity : AppCompatActivity() {
+class MyReviewActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_user_modify)
+        setContentView(R.layout.activity_my_review)
 
         btn_back.setOnClickListener {
             finish()
