+++ app/src/main/java/com/yeonae/chamelezone/view/place/PlaceDetailActivity.kt
@@ -1,8 +1,11 @@
-package com.yeonae.chamelezone
+package com.yeonae.chamelezone.view.place
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.yeonae.chamelezone.R
+import com.yeonae.chamelezone.view.review.ReviewCreateActivity
+import com.yeonae.chamelezone.adapter.ImageViewPagerAdapter
 import kotlinx.android.synthetic.main.activity_place_detail.*
 //import me.relex.circleindicator.CircleIndicator
