+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context=".HomeActivity">
+    tools:context=".view.home.HomeActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
